ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
numpy==1.24.4
gensim==4.3.2
mysql-connector-python==8.4.0
konlpy==0.6.0
aiofiles==23.2.1
scikit-learn==1.3.2